Add extra buttons.
No simple way to control the handoff of control between main thread and
worker generator. This makes it hard to predict without starting
training first, or continuing training after prediction.

Instead have a predict button as part of training that will only predict
on the raw training data as a whole for visualization purposes.

Predicting on new data will be limited to the prediction dropdown