Breaking change: Add support for lsds/fgbg
Adds support for training/predicting fgbg and lsds with affinities.
We need a way to determine which output is fgbg/affinities/lsds. I chose
to identify them via name here. Is that the best option?
To get this to work with the existing model, you need to change the
name of the output tensor of the sample model from output0 to affinities