Remove affogato as a dependency.
also update README with updated installation instructions.
affogato was removed as a python dependency because there is a different
package on pypi also called affogato that we do not want. This is also
breaking the conda recipe tests since pip check does not recognize the
conda version of affogato as satisfying the dependency